test: track coverage for plugins
