--- .github/workflows/plugin_test.yml
@@ -1,24 +1,0 @@
-name: Plugins Test
-
-on:
-  push:
-    branches:
-      - main
-  pull_request:
-  workflow_dispatch:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v3
-      - name: Set up Go
-        uses: actions/setup-go@v3
-        with:
-          go-version: '1.18'
-          cache: true
-      - name: Install dependencies
-        run: sudo apt-get install build-essential
-      - name: Run Test
-        run: make test-plugins
