+++ .github/workflows/test.yml
@@ -1,5 +1,9 @@
 name: Test
-on: [push, pull_request]
+
+on:
+  push:
+  pull_request:
+  workflow_dispatch:
 
 jobs:
   test:
@@ -16,9 +20,58 @@ jobs:
         run: sudo apt-get install build-essential
       - name: Run Test
         run: make test
+      - name: Upload coverage artifact
+        uses: actions/upload-artifact@v3
+        with:
+          name: coverage
+          path: coverage.out
+
+  plugins-test:
+    runs-on: ubuntu-latest
+    if: |
+      github.ref == 'refs/heads/main' || 
+      github.event_name == 'pull_request' ||
+      github.event_name == 'workflow_dispatch'
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+      - name: Set up Go
+        uses: actions/setup-go@v3
+        with:
+          go-version: '1.18'
+          cache: true
+      - name: Install dependencies
+        run: sudo apt-get install build-essential
+      - name: Run Test
+        run: make test-plugins
+      - name: Upload coverage artifact
+        uses: actions/upload-artifact@v3
+        with:
+          name: coverage-plugins
+          path: coverage-plugins.out
+
+  coverage:
+    runs-on: ubuntu-latest
+    needs: [test, plugins-test]
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v3
+      - name: Set up Go
+        uses: actions/setup-go@v3
+        with:
+          go-version: '1.18'
+          cache: true
+      - name: Download coverage
+        uses: actions/download-artifact@v3
+        with:
+          name: coverage
+      - name: Download plugins coverage
+        uses: actions/download-artifact@v3
+        with:
+          name: coverage-plugins
       - name: Install goveralls and send coverage
         env:
           COVERALLS_TOKEN: ${{ secrets.GITHUB_TOKEN }}
         run: |
           go install github.com/mattn/goveralls@v0.0.11
-          goveralls -coverprofile=coverage.out -service=github
+          goveralls -coverprofile=coverage.out,coverage-plugins.out -service=github
