+++ plugins/extractors/bigquery/auditlog/testdata_test.go
@@ -1,3 +1,6 @@
+//go:build plugins
+// +build plugins
+
 package auditlog
 
 import (
